modify file names to use snake case
