+++ core/rammp_prototype_behavior/rammp_prototype_behavior/actionClient/chair_crub_traverse_action_client.py
@@ -1,19 +1,25 @@
 from __future__ import annotations
 import asyncio
+import enum
 
 from rclpy.node import Node
-from cornell_feeding_interfaces.action import CornellActionsPlaceHolder
-from .ActionClientWrapper import ActionClientWrapper
+from rammp_prototype_interfaces.action import CurbTraverse
+from .action_client_wrapper import ActionClientWrapper
 
 
-class BringCupToMouthActionClient(ActionClientWrapper):
+class CurbTraverseDirection(enum.IntEnum):
+    ASCEND = CurbTraverse.Goal.ASCEND
+    DESCEND = CurbTraverse.Goal.DESCEND
+
+
+class ChairCurbTraverseActionClient(ActionClientWrapper):
     def __init__(
         self,
         node: Node,
     ):
         super().__init__(
-            "/arm/drink/BringCupToMouth",
-            CornellActionsPlaceHolder,
+            "/base/curb_traverse",
+            CurbTraverse,
             self.goal_callback,
             self.result_callback,
             self.cancel_callback,
@@ -23,36 +29,38 @@ class BringCupToMouthActionClient(ActionClientWrapper):
     def goal_callback(self, success: bool):
         if success:
             self._node.get_logger().info(
-                "Goal BringCupToMouth accepted by the action server."
+                "Goal CurbTraverse accepted by the action server."
             )
         else:
             self._node.get_logger().warn(
-                "Goal BringCupToMouth rejected by the action server."
+                "Goal CurbTraverse rejected by the action server."
             )
-            self._node.reqArmActionGoalFailed()
+            self._node.reqNavCancel()
 
     def result_callback(self, success: bool):
         if success:
-            self._node.get_logger().info("Successfully brought cup to mouth.")
-            self._node.readyForDrink()  # should enter readyForDrink state after bringing cup to mouth
+            self._node.get_logger().info("Successfully completed CurbTraverse.")
+            self._node.traverseComplete()
+            self._node.finish_mock_task()  # for testing, will remove after testing
         else:
-            self._node.get_logger().warn("Failed to bring cup to mouth.")
-            self._node.ArmActionFailed()
+            self._node.get_logger().warn("Failed to complete CurbTraverse.")
+            self._node.reqNavCancel()
 
     def cancel_callback(self, success: bool):
         if success:
             self._node.get_logger().info(
                 "Goal cancellation accepted by the action server."
             )
-            self._node.reqArmActionCancelSuccess()
+            self._node.reqNavCancel()
         else:
             self._node.get_logger().warn(
                 "Goal cancellation rejected by the action server."
             )
-            self._node.reqArmActionCancelFailed()
+            self._node.reqNavCancel()
 
-    def send_goal(self):
-        goal = CornellActionsPlaceHolder.Goal()
+    def send_goal(self, direction: CurbTraverseDirection):
+        goal = CurbTraverse.Goal()
+        goal.direction = direction.value
         asyncio.run_coroutine_threadsafe(super().send_goal(goal), self._node._loop)
 
     def cancel(self):
