+++ core/rammp_prototype_behavior/rammp_prototype_behavior/actionClient/open_door_action_client.py
@@ -2,18 +2,18 @@ from __future__ import annotations
 import asyncio
 
 from rclpy.node import Node
-from cornell_feeding_interfaces.action import CornellActionsPlaceHolder
-from .ActionClientWrapper import ActionClientWrapper
+from cmu_door_opener_interfaces.action import DoorOpenActionTypePlaceHolder
+from .action_client_wrapper import ActionClientWrapper
 
 
-class PutCupBackToHolderActionClient(ActionClientWrapper):
+class OpenDoorActionClient(ActionClientWrapper):
     def __init__(
         self,
         node: Node,
     ):
         super().__init__(
-            "/arm/drink/PutCupBackToHolder",
-            CornellActionsPlaceHolder,
+            "/arm/door/open",
+            DoorOpenActionTypePlaceHolder,
             self.goal_callback,
             self.result_callback,
             self.cancel_callback,
@@ -22,23 +22,19 @@ class PutCupBackToHolderActionClient(ActionClientWrapper):
 
     def goal_callback(self, success: bool):
         if success:
-            self._node.get_logger().info(
-                "Goal PutCupBackToHolder accepted by the action server."
-            )
+            self._node.get_logger().info("Goal OpenDoor accepted by the action server.")
         else:
-            self._node.get_logger().warn(
-                "Goal PutCupBackToHolder rejected by the action server."
-            )
+            self._node.get_logger().warn("Goal OpenDoor rejected by the action server.")
             self._node.reqArmActionGoalFailed()
 
     def result_callback(self, success: bool):
         if success:
-            self._node.get_logger().info("Successfully put cup back to holder.")
-            self._node.cupPlacedBack()
-            self._node.set_arm_mode_idle()  # set arm to idle after putting cup back to holder
+            self._node.get_logger().info("Successfully opened door.")
+            self._node.doorOpenFinished()
+            self._node.set_arm_mode_idle()  # set arm to idle after opening door
             self._node.finish_mock_task()  # for testing, will remove after testing
         else:
-            self._node.get_logger().warn("Failed to put cup back to holder.")
+            self._node.get_logger().warn("Failed to open door.")
             self._node.ArmActionFailed()
 
     def cancel_callback(self, success: bool):
@@ -54,7 +50,7 @@ class PutCupBackToHolderActionClient(ActionClientWrapper):
             self._node.reqArmActionCancelFailed()
 
     def send_goal(self):
-        goal = CornellActionsPlaceHolder.Goal()
+        goal = DoorOpenActionTypePlaceHolder.Goal()
         asyncio.run_coroutine_threadsafe(super().send_goal(goal), self._node._loop)
 
     def cancel(self):
