+++ core/rammp_prototype_behavior/rammp_prototype_behavior/actionClient/pick_up_and_order_action_client.py
@@ -3,16 +3,16 @@ import asyncio
 
 from rclpy.node import Node
 from cornell_feeding_interfaces.action import CornellActionsPlaceHolder
-from .ActionClientWrapper import ActionClientWrapper
+from .action_client_wrapper import ActionClientWrapper
 
 
-class HomeCupActionClient(ActionClientWrapper):
+class PickUpAndOrderActionClient(ActionClientWrapper):
     def __init__(
         self,
         node: Node,
     ):
         super().__init__(
-            "/arm/drink/HomeCup",
+            "/arm/drink/PickupAndOrder",
             CornellActionsPlaceHolder,
             self.goal_callback,
             self.result_callback,
@@ -22,19 +22,21 @@ class HomeCupActionClient(ActionClientWrapper):
 
     def goal_callback(self, success: bool):
         if success:
-            self._node.get_logger().info("Goal HomeCup accepted by the action server.")
+            self._node.get_logger().info(
+                "Goal PickUpAndOrder accepted by the action server."
+            )
         else:
-            self._node.get_logger().warn("Goal HomeCup rejected by the action server.")
+            self._node.get_logger().warn(
+                "Goal PickUpAndOrder rejected by the action server."
+            )
             self._node.reqArmActionGoalFailed()
 
     def result_callback(self, success: bool):
         if success:
-            self._node.get_logger().info("Successfully homed cup.")
-            self._node.homedCup()  # should enter homed state after homing cup
-            self._node.set_arm_mode_idle()  # set arm to idle after homing cup
-            self._node.finish_mock_task()  # for testing, will remove after testing
+            self._node.get_logger().info("Successfully picked up and ordered.")
+            self._node.pickedUpCup()
         else:
-            self._node.get_logger().warn("Failed to home cup.")
+            self._node.get_logger().warn("Failed to pick up and order cup.")
             self._node.ArmActionFailed()
 
     def cancel_callback(self, success: bool):
