+++ core/rammp_prototype_behavior/rammp_prototype_behavior/actionClient/pub_cup_back_to_holder_action_client.py
@@ -1,25 +1,19 @@
 from __future__ import annotations
 import asyncio
-import enum
 
 from rclpy.node import Node
-from rammp_prototype_interfaces.action import CurbTraverse
-from .ActionClientWrapper import ActionClientWrapper
+from cornell_feeding_interfaces.action import CornellActionsPlaceHolder
+from .action_client_wrapper import ActionClientWrapper
 
 
-class CurbTraverseDirection(enum.IntEnum):
-    ASCEND = CurbTraverse.Goal.ASCEND
-    DESCEND = CurbTraverse.Goal.DESCEND
-
-
-class ChairCurbTraverseActionClient(ActionClientWrapper):
+class PutCupBackToHolderActionClient(ActionClientWrapper):
     def __init__(
         self,
         node: Node,
     ):
         super().__init__(
-            "/base/curb_traverse",
-            CurbTraverse,
+            "/arm/drink/PutCupBackToHolder",
+            CornellActionsPlaceHolder,
             self.goal_callback,
             self.result_callback,
             self.cancel_callback,
@@ -29,38 +23,38 @@ class ChairCurbTraverseActionClient(ActionClientWrapper):
     def goal_callback(self, success: bool):
         if success:
             self._node.get_logger().info(
-                "Goal CurbTraverse accepted by the action server."
+                "Goal PutCupBackToHolder accepted by the action server."
             )
         else:
             self._node.get_logger().warn(
-                "Goal CurbTraverse rejected by the action server."
+                "Goal PutCupBackToHolder rejected by the action server."
             )
-            self._node.reqNavCancel()
+            self._node.reqArmActionGoalFailed()
 
     def result_callback(self, success: bool):
         if success:
-            self._node.get_logger().info("Successfully completed CurbTraverse.")
-            self._node.traverseComplete()
+            self._node.get_logger().info("Successfully put cup back to holder.")
+            self._node.cupPlacedBack()
+            self._node.set_arm_mode_idle()  # set arm to idle after putting cup back to holder
             self._node.finish_mock_task()  # for testing, will remove after testing
         else:
-            self._node.get_logger().warn("Failed to complete CurbTraverse.")
-            self._node.reqNavCancel()
+            self._node.get_logger().warn("Failed to put cup back to holder.")
+            self._node.ArmActionFailed()
 
     def cancel_callback(self, success: bool):
         if success:
             self._node.get_logger().info(
                 "Goal cancellation accepted by the action server."
             )
-            self._node.reqNavCancel()
+            self._node.reqArmActionCancelSuccess()
         else:
             self._node.get_logger().warn(
                 "Goal cancellation rejected by the action server."
             )
-            self._node.reqNavCancel()
+            self._node.reqArmActionCancelFailed()
 
-    def send_goal(self, direction: CurbTraverseDirection):
-        goal = CurbTraverse.Goal()
-        goal.direction = direction.value
+    def send_goal(self):
+        goal = CornellActionsPlaceHolder.Goal()
         asyncio.run_coroutine_threadsafe(super().send_goal(goal), self._node._loop)
 
     def cancel(self):
