+++ core/rammp_prototype_behavior/setup.py
@@ -29,7 +29,7 @@ setup(
             "test_nn_monitor = rammp_prototype_behavior.test.test_nn_monitor:main",  # just for testing
             "test_node1 = rammp_prototype_behavior.test.test_node1:main",  # just for testing
             "test_node2 = rammp_prototype_behavior.test.test_node2:main",  # just for testing
-            "system_control = rammp_prototype_behavior.SystemControl:main",
+            "system_control = rammp_prototype_behavior.system_control:main",
             "mock_arm_driver = rammp_prototype_behavior.mocks.mock_arm_driver:main",
             "mock_drinking_node = rammp_prototype_behavior.mocks.mock_drinking_node:main",
             "mock_opening_door = rammp_prototype_behavior.mocks.mock_opening_door:main",
